loading in multiple files and working toward stored calculations
added a jupyter notebook for working on preloading calculations

added flsyst module that serves to load in data and provide vectors for indexing through data based on the subject, angle, eccentricity, conetype (when relevant), and filetype.  the main function to use from flsyst is getConeData() which returns all files described by the users category inputs and index vectors for each category as well as an index vector that labels data by mosaic (as each mosaic has many files associated with it).

getConeData relies on 4 other functions in the filesyst module: 1) getFileStrings(), 2) getIndices(), 3), getFileVectors() 4) getFilesByDataGroup()

Also added show and calc modules and divvied up the functions previously housed in utilities between them, and added a function for plotting the cone mosaic on the ROI image to show.