LARGE COMMIT KRAKEN TRANSITION
had to reverse two commits that were already too large for their own good.  added flake8 and h5py to the environment, massively edited the modules so far to be flake8 formatted, updated the filesystem so that it can only take one filetype at a time (much simpler for all indexing needs), added a filesyst function that reorganizes data to be compartmentalized by which dataset it's from, and added many optional inputs to plotting functions to make them easier to use.  Also changed things to save off of the git repo but left the original dataset there because it is small. Installed Git so that I could reverse those commits (contained huge files) from bash, and Git Kraken for future easier GUI use.