preloaded data strides
settled on organization for data to be saved (per data file, all the data processed from that file is saved in one hdF5).  fixed issues with my distance calculating function and added flexibility.  successfully saving files and now working on a flexible operation for only running calculations if they've not already been saved in the folder.  will need to add an override option.