avg nearest cone dist in plots, save plots
added stuff to get nearest cone dist and plot it in MCU normed plots.   that is not the place for that operation to occur.  but it is what's happening for now bc in a hurry.  added some parameters and lines to all the smp view scripts to be able to save them and added save directory names and directory making lines to the View SMP jupyter notebook, but haven't actually tested any of it.  would rather do that on the other computer which has been generating the last batch of simulations meanwhile.