manually transferred spacified code
git issues on lab computer, manually copied changes over to laptop.  added functionality for a spacified process to single mosaic processes, savefl_config, the smp jupyter notebook, and calc.  also fixed a bug in the single mosaic process for viewing MCU normed stuff that was skipping a lot of files.