changing to 'simulations' and 'analyses'
starting a big change to have simulations and analysis.  this in one swoop will knock out a ton of the redundancy in the SMP code and build in the needed flexibility for secondary and tertiary analyses, like DRP and 2PC (secondary) and post-2PC metrics (tertiary).  wrote some notes about what makes something primary and tertiary in lab notebook and the logic behind the simulation/analysis hierarchy.  updated a lot of the configuration file and smp code, nothing is functional yet.  still need to add the content of the hierarchy layer processes and start testing things to learn about all the fun things i didn't consider