new 2PC function running/saving knockonwood
changing processes to simulations and analyses took a LOTALOT of work.  altered the whole flow of the pipeline and the system by which data is referenced and indexed. 

made intracone_distance the sole primary function and two_point_correlation the sole secondary function.  

not a huge fan of the separate functions for "primary_processes" and "secondary_processes" and these being separate groups in the save configuration file. this can be simplified into subgroups in the savefile i'm sure. perhaps default processes can also be incorporated into that. 

got the desired flexibility for running the 2PC function.  have not caught up to visualizing it so am not sure how well it's actually working, just that it seems to be saving reasonable enough looking data. 

fixed the visualizer for spacified mosaics.  all the other visualizers still need modified for the new indexing system, new names, and 2PC flexibility

cheers, this was a big step.  now debug and test the butts off of it, then this branch will be done.