nearest neighbor spacify flexibility
added flexibility to the spacify by nearest neighbors process to take an input of how many neighbors are arveraged. also added an option for summing instead of averaging, that I haven't played with much yet.  Added some functionality to flsyst to be able to handle empty list inputs for sim_to_gen and analyses_to_run