edited basic stats, added hexgrid_by_density
made basic stats a default process instead of a primary analysis, and made it so that for every process it's run on the main file than on each of the simulated files.  added hexgrid_by_density function for generating hexagonal mosaics based on given cone density.  not perfect yet, it's not set within the image bounds. but it's an ok start.