fixed the scrunched x-axis
found what was causing the bug that led all the 2PC figures to be scrunched1/2 as big on the x axis as they should have been.  It was in the plotting, in a line where I set the x values of the histogram to be the center of the bin rather than the edge. I was halving all the bin edges rather than subtracting half the binwidth from them. fixed!