test hexgrid data, voronoi process, some unit conversion
added hexgrid data for two hexagonal point distributions with the same number of points with different densities to test density dependence of analyses

added voronoi process (a lot of copied code) for calculating and drawing voronoi diagrams. identifies bound voronoi cells (buffer needs to be set higher for real data it seems, I don't totally follow how the voronoi facets are being set near the edges), then calculates voronoi cell areas & number of neighbors and bound cell density.  Added functions for viewing the voronoi cells colored by number of neighbors and bound status, and to plot the metric histograms of bound voronoi cells

added some conversion flexibility. can convert the coord (and img) unit by a conversion factor, and can set the density to be converted (this is useful bc arcsec resolution is good for cv2's functionality, but shitty for talking about density which we tend to prefer in degrees. this could use some work in the future but it's nice for now.  There's probably still some places in the code that are still reading in the image size from the image directly, which needs to be fixed to read from the img_x and img_y variables which will be based on the conversions if any.