view voronoi function
added view voronoi functions, fixed skipping of saving mean and std of voronoi metrics if the std was 0, only mean to be skipping the regularity if std was 0 to avoid divide-by-zero issues.  Next need to add in the last of the voronoi functionality (density) and make sure that it works with unit conversions.