'getAll', subsample data, Roorda data,  binary distributions
Got as far with adding the 'getAll' functionality for input config files as making sure it works for pulling all subjects in a file (what I needed it to do for subsampling analysis, where there are bazillions of "subjects"). Haven't tested it with angle or eccentricity or conetype (though intuitively it shouldn't be any different for those variables), nor with multiple values set to 'getAll' at a time.

Replaced subsampled data from subject 1 1.5 deg temporal (example ROI) with subject 2 1.5 deg nasal (ROI with the most cones). Added data from Roorda 2001

Added code to calc that generates binary distributions of S-cone occurrence as a function of distance from other S-cones.